Checkpoint commit, still trying to get tests working

+++ core/src/test/java/com/datastax/oss/driver/internal/core/type/codec/registry/CachingCodecRegistryTestDataProviders.java
@@ -296,43 +296,43 @@ public class CachingCodecRegistryTestDataProviders {
       {
         DataTypes.vectorOf(DataTypes.BIGINT, 1),
         GenericType.vectorOf(Long.class),
-        GenericType.vectorOf(Long.class),
+        GenericType.vectorOf(Integer.class),
         CqlVector.newInstance(1l)
       },
       {
         DataTypes.vectorOf(DataTypes.SMALLINT, 1),
         GenericType.vectorOf(Short.class),
-        GenericType.vectorOf(Short.class),
-        CqlVector.newInstance(1)
+        GenericType.vectorOf(Integer.class),
+        CqlVector.newInstance((short) 1)
       },
       {
         DataTypes.vectorOf(DataTypes.TINYINT, 1),
         GenericType.vectorOf(Byte.class),
-        GenericType.vectorOf(Byte.class),
-        CqlVector.newInstance(1)
+        GenericType.vectorOf(Integer.class),
+        CqlVector.newInstance((byte) 1)
       },
       {
         DataTypes.vectorOf(DataTypes.FLOAT, 1),
         GenericType.vectorOf(Float.class),
-        GenericType.vectorOf(Float.class),
+        GenericType.vectorOf(Integer.class),
         CqlVector.newInstance(1.0f)
       },
       {
         DataTypes.vectorOf(DataTypes.DOUBLE, 1),
         GenericType.vectorOf(Double.class),
-        GenericType.vectorOf(Double.class),
+        GenericType.vectorOf(Integer.class),
         CqlVector.newInstance(1.0d)
       },
       {
         DataTypes.vectorOf(DataTypes.DECIMAL, 1),
         GenericType.vectorOf(BigDecimal.class),
-        GenericType.vectorOf(BigDecimal.class),
+        GenericType.vectorOf(Integer.class),
         CqlVector.newInstance(BigDecimal.ONE)
       },
       {
         DataTypes.vectorOf(DataTypes.VARINT, 1),
         GenericType.vectorOf(BigInteger.class),
-        GenericType.vectorOf(BigInteger.class),
+        GenericType.vectorOf(Integer.class),
         CqlVector.newInstance(BigInteger.ONE)
       },
     };
